+++ graphics/java/android/graphics/DashPathEffect.java
@@ -38,6 +38,6 @@ public class DashPathEffect extends PathEffect {
         native_instance = nativeCreate(intervals, phase);
     }
     
-    private static native int nativeCreate(float intervals[], float phase);
+    private static native long nativeCreate(float intervals[], float phase);
 }
